requirements.txt
ultralytics>=8.2.0
opencv-python
numpy
pymavlink
pytz
PyYAML
python-dotenv